Add routes and create route for spotify
